Remove unused files and update project configurations for Flutter and Dart compatibility. Added new settings and improved code structure across various screens and providers.

+++ lib/plugins/geographic_msgs/lib/src/msg/BoundingBox.dart
@@ -25,7 +25,7 @@ GeoPoint max_pt''';
 
   @override
   int getMessageSize() {
-    return (min_pt.getMessageSize() as int) + (max_pt.getMessageSize() as int);
+    return (min_pt.getMessageSize()) + (max_pt.getMessageSize());
   }
 
   @override
