+++ lib/plugins/geographic_msgs/lib/src/msg/GeoPoseStamped.dart
@@ -26,7 +26,7 @@ geographic_msgs/GeoPose pose''';
 
   @override
   int getMessageSize() {
-    return (header.getMessageSize() as int) + (pose.getMessageSize() as int);
+    return (header.getMessageSize()) + (pose.getMessageSize());
   }
 
   @override
