+++ lib/plugins/nav2_msgs/lib/src/action/ComputePathToPose.dart
@@ -38,8 +38,8 @@ bool use_start''';
 
   @override
   int getMessageSize() {
-    return (goal.getMessageSize() as int) +
-        (start.getMessageSize() as int) +
+    return (goal.getMessageSize()) +
+        (start.getMessageSize()) +
         4 +
         planner_id.length +
         1;
@@ -93,9 +93,7 @@ uint16 error_code''';
 
   @override
   int getMessageSize() {
-    return (path.getMessageSize() as int) +
-        (planning_time.getMessageSize() as int) +
-        2;
+    return (path.getMessageSize()) + (planning_time.getMessageSize()) + 2;
   }
 
   @override
