+++ lib/plugins/nav2_msgs/lib/src/action/DriveOnHeading.dart
@@ -33,9 +33,7 @@ builtin_interfaces/Duration time_allowance''';
 
   @override
   int getMessageSize() {
-    return (target.getMessageSize() as int) +
-        4 +
-        (time_allowance.getMessageSize() as int);
+    return (target.getMessageSize()) + 4 + (time_allowance.getMessageSize());
   }
 
   @override
@@ -81,7 +79,7 @@ uint16 error_code''';
 
   @override
   int getMessageSize() {
-    return (total_elapsed_time.getMessageSize() as int) + 2;
+    return (total_elapsed_time.getMessageSize()) + 2;
   }
 
   @override
